Move the examples to the lab, as we need to write actual tutorials
