+++ Silk.NET.sln
@@ -15,7 +15,7 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Silk.NET.Windowing.Common",
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Silk.NET.Windowing.Desktop", "src\Windowing\Silk.NET.Windowing.Desktop\Silk.NET.Windowing.Desktop.csproj", "{0A18FCAE-572E-47FF-B8E3-C97ED15132FA}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BlankWindow", "examples\BlankWindow\BlankWindow.csproj", "{E8C4587E-9424-438B-AEAF-A09E0234865A}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BlankWindow", "src\Lab\BlankWindow\BlankWindow.csproj", "{E8C4587E-9424-438B-AEAF-A09E0234865A}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Silk.NET.Windowing", "src\Windowing\Silk.NET.Windowing\Silk.NET.Windowing.csproj", "{956F722C-DFD3-435E-9D2E-A0549C4D8BC0}"
 EndProject
@@ -165,23 +165,23 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Silk.NET.OpenGLES.Extension
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Core", "Core", "{0651C5EF-50AA-4598-8D9C-8F210ADD8490}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Triangle", "examples\Triangle\Triangle.csproj", "{3478F392-7055-4ECC-B388-7F115AA7EA5A}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Triangle", "src\Lab\Triangle\Triangle.csproj", "{3478F392-7055-4ECC-B388-7F115AA7EA5A}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Silk.NET.Input.Desktop", "src\Input\Silk.NET.Input.Desktop\Silk.NET.Input.Desktop.csproj", "{3908DEF6-7403-49F5-B8EC-5B3B12C325D4}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Silk.NET.Input", "src\Input\Silk.NET.Input\Silk.NET.Input.csproj", "{020A8E88-B607-4281-BA0D-5ED03484A201}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "InputTest", "examples\InputTest\InputTest.csproj", "{3E744E54-F450-4051-8919-04D9B591688E}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "InputTest", "src\Lab\InputTest\InputTest.csproj", "{3E744E54-F450-4051-8919-04D9B591688E}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Silk.NET", "src\Core\Silk.NET\Silk.NET.csproj", "{7BB1DE56-DA1C-4776-A961-B810574E1EF4}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SampleBase", "examples\SampleBase\SampleBase.csproj", "{D702E338-BCB7-432C-8896-6B6086821B3E}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SampleBase", "src\Lab\SampleBase\SampleBase.csproj", "{D702E338-BCB7-432C-8896-6B6086821B3E}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Vulkan", "Vulkan", "{E2ABDF45-C329-47B2-8E09-B7298E2557F7}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Silk.NET.Vulkan", "src\Vulkan\Silk.NET.Vulkan\Silk.NET.Vulkan.csproj", "{7F347EDA-EC20-4E3F-9C25-CF32ED7D9642}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "VulkanTriangle", "examples\VulkanTriangle\VulkanTriangle.csproj", "{AC1B288B-2F07-49EB-8518-9DA27C6D6503}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "VulkanTriangle", "src\Lab\VulkanTriangle\VulkanTriangle.csproj", "{AC1B288B-2F07-49EB-8518-9DA27C6D6503}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Extensions", "Extensions", "{49BC383A-D7E6-4013-93C7-371479B984CC}"
 EndProject
@@ -210,6 +210,12 @@ EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Silk.NET.Vulkan.Extensions.AMD", "src\Vulkan\Extensions\Silk.NET.Vulkan.Extensions.AMD\Silk.NET.Vulkan.Extensions.AMD.csproj", "{3D945DE0-1E2D-439E-991D-9C65C325F82F}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Silk.NET.OpenGL.Extensions.MESA", "src\OpenGL\Extensions\Silk.NET.OpenGL.Extensions.MESA\Silk.NET.OpenGL.Extensions.MESA.csproj", "{D27D10F0-F22B-49B5-9933-375BB120AD70}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Lab", "Lab", "{DFA0E841-33E5-4533-AF00-964E21A141B8}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "C#", "C#", "{974C560A-E6FE-40BF-B1DF-24AEACBA34A9}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "F#", "F#", "{725B09E8-CD6A-4DA8-8686-F2ACA9D10FFB}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -1334,7 +1340,6 @@ Global
 		{A8DFEA18-E880-4303-BBCE-0F27EE214D44} = {23324041-2076-477C-A4BF-B385B8066C6C}
 		{0A18FCAE-572E-47FF-B8E3-C97ED15132FA} = {23324041-2076-477C-A4BF-B385B8066C6C}
 		{956F722C-DFD3-435E-9D2E-A0549C4D8BC0} = {23324041-2076-477C-A4BF-B385B8066C6C}
-		{E8C4587E-9424-438B-AEAF-A09E0234865A} = {E1F91563-7277-4E9B-A3B7-8D5FD9802A4A}
 		{5A41E9D5-67F5-447C-8AE2-78FED7A45206} = {081E7761-B200-4DBF-8950-941464DECACE}
 		{896B1A79-FEC8-4B7A-8A9C-0CC8EDA2CCFC} = {081E7761-B200-4DBF-8950-941464DECACE}
 		{FC015604-3326-4BCC-9212-C5CB0DFB0D12} = {896B1A79-FEC8-4B7A-8A9C-0CC8EDA2CCFC}
@@ -1402,14 +1407,10 @@ Global
 		{8D534FE1-6113-457A-97DD-FD42782580A5} = {CF69D5C3-4ACE-4458-BA5A-0E9A3B294CDC}
 		{FD24E9FF-1097-4777-A418-F2D88C558665} = {0651C5EF-50AA-4598-8D9C-8F210ADD8490}
 		{136C6154-D300-4B82-80D3-17B637841A2B} = {0651C5EF-50AA-4598-8D9C-8F210ADD8490}
-		{3478F392-7055-4ECC-B388-7F115AA7EA5A} = {E1F91563-7277-4E9B-A3B7-8D5FD9802A4A}
 		{3908DEF6-7403-49F5-B8EC-5B3B12C325D4} = {FA9D1C95-5585-4DEC-B226-1447A486C376}
 		{020A8E88-B607-4281-BA0D-5ED03484A201} = {FA9D1C95-5585-4DEC-B226-1447A486C376}
-		{3E744E54-F450-4051-8919-04D9B591688E} = {E1F91563-7277-4E9B-A3B7-8D5FD9802A4A}
 		{7BB1DE56-DA1C-4776-A961-B810574E1EF4} = {0651C5EF-50AA-4598-8D9C-8F210ADD8490}
-		{D702E338-BCB7-432C-8896-6B6086821B3E} = {E1F91563-7277-4E9B-A3B7-8D5FD9802A4A}
 		{7F347EDA-EC20-4E3F-9C25-CF32ED7D9642} = {E2ABDF45-C329-47B2-8E09-B7298E2557F7}
-		{AC1B288B-2F07-49EB-8518-9DA27C6D6503} = {E1F91563-7277-4E9B-A3B7-8D5FD9802A4A}
 		{49BC383A-D7E6-4013-93C7-371479B984CC} = {E2ABDF45-C329-47B2-8E09-B7298E2557F7}
 		{043852EA-FBD4-4F42-9CDC-92078F9EF942} = {49BC383A-D7E6-4013-93C7-371479B984CC}
 		{D860135C-01A2-4597-B1B2-DA4882C09798} = {49BC383A-D7E6-4013-93C7-371479B984CC}
@@ -1424,5 +1425,12 @@ Global
 		{65E16908-BD7A-4FC2-BC23-FD8194039C5B} = {49BC383A-D7E6-4013-93C7-371479B984CC}
 		{3D945DE0-1E2D-439E-991D-9C65C325F82F} = {49BC383A-D7E6-4013-93C7-371479B984CC}
 		{D27D10F0-F22B-49B5-9933-375BB120AD70} = {CF69D5C3-4ACE-4458-BA5A-0E9A3B294CDC}
+		{E8C4587E-9424-438B-AEAF-A09E0234865A} = {DFA0E841-33E5-4533-AF00-964E21A141B8}
+		{3E744E54-F450-4051-8919-04D9B591688E} = {DFA0E841-33E5-4533-AF00-964E21A141B8}
+		{D702E338-BCB7-432C-8896-6B6086821B3E} = {DFA0E841-33E5-4533-AF00-964E21A141B8}
+		{3478F392-7055-4ECC-B388-7F115AA7EA5A} = {DFA0E841-33E5-4533-AF00-964E21A141B8}
+		{AC1B288B-2F07-49EB-8518-9DA27C6D6503} = {DFA0E841-33E5-4533-AF00-964E21A141B8}
+		{974C560A-E6FE-40BF-B1DF-24AEACBA34A9} = {E1F91563-7277-4E9B-A3B7-8D5FD9802A4A}
+		{725B09E8-CD6A-4DA8-8686-F2ACA9D10FFB} = {E1F91563-7277-4E9B-A3B7-8D5FD9802A4A}
 	EndGlobalSection
 EndGlobal
