--- examples/VulkanTriangle/VulkanTriangle.csproj
@@ -1,24 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-    <PropertyGroup>
-        <OutputType>Exe</OutputType>
-        <TargetFramework>netcoreapp3.0</TargetFramework>
-        <AllowUnsafeBlocks>true</AllowUnsafeBlocks>
-    </PropertyGroup>
-
-    <ItemGroup>
-      <ProjectReference Include="..\..\src\Vulkan\Extensions\Silk.NET.Vulkan.Extensions.EXT\Silk.NET.Vulkan.Extensions.EXT.csproj" />
-      <ProjectReference Include="..\..\src\Vulkan\Extensions\Silk.NET.Vulkan.Extensions.KHR\Silk.NET.Vulkan.Extensions.KHR.csproj" />
-      <ProjectReference Include="..\..\src\Vulkan\Silk.NET.Vulkan\Silk.NET.Vulkan.csproj" />
-      <ProjectReference Include="..\..\src\Windowing\Silk.NET.Windowing.Common\Silk.NET.Windowing.Common.csproj" />
-      <ProjectReference Include="..\..\src\Windowing\Silk.NET.Windowing\Silk.NET.Windowing.csproj" />
-    </ItemGroup>
-
-    <ItemGroup>
-      <None Remove="shader.frag.spv" />
-      <EmbeddedResource Include="shader.frag.spv" />
-      <None Remove="shader.vert.spv" />
-      <EmbeddedResource Include="shader.vert.spv" />
-    </ItemGroup>
-
-</Project>
