+++ src/Lab/BlankWindow/BlankWindow.csproj
@@ -6,9 +6,9 @@
     </PropertyGroup>
 
     <ItemGroup>
-      <ProjectReference Include="..\..\src\Windowing\Silk.NET.Windowing\Silk.NET.Windowing.csproj" />
+      <ProjectReference Include="..\..\Windowing\Silk.NET.Windowing\Silk.NET.Windowing.csproj" />
     </ItemGroup>
 
-    <Import Project="..\..\build\props\common.props" />
+    <Import Project="..\..\..\build\props\common.props" />
 
 </Project>
