+++ src/Lab/InputTest/Program.cs
