+++ src/Lab/Triangle/shader.frag
