+++ src/Lab/VulkanTriangle/shader.frag.spv
