+++ src/Lab/VulkanTriangle/shader.vert.spv
